Move browser code to separate file.
* Move to browser package so the source from Go authors does not need to be
  changed.
