+++ contrib/jepsen/main.go
@@ -36,10 +36,11 @@ import (
 	"log"
 	"os"
 	"os/exec"
-	"runtime"
 	"strconv"
 	"strings"
 	"time"
+
+	"github.com/dgraph-io/dgraph/contrib/jepsen/browser"
 )
 
 type JepsenTest struct {
@@ -178,63 +179,7 @@ func openJepsenBrowser() {
 	}
 	port := strings.TrimSpace(out.String())
 	jepsenUrl := "http://localhost:" + port
-	BrowserOpen(jepsenUrl)
-}
-
-// Copied from $GOROOT/src/cmd/internal/browser/browser.go
-
-// Commands returns a list of possible commands to use to open a url.
-func BrowserCommands() [][]string {
-	var cmds [][]string
-	if exe := os.Getenv("BROWSER"); exe != "" {
-		cmds = append(cmds, []string{exe})
-	}
-	switch runtime.GOOS {
-	case "darwin":
-		cmds = append(cmds, []string{"/usr/bin/open"})
-	case "windows":
-		cmds = append(cmds, []string{"cmd", "/c", "start"})
-	default:
-		if os.Getenv("DISPLAY") != "" {
-			// xdg-open is only for use in a desktop environment.
-			cmds = append(cmds, []string{"xdg-open"})
-		}
-	}
-	cmds = append(cmds,
-		[]string{"chrome"},
-		[]string{"google-chrome"},
-		[]string{"chromium"},
-		[]string{"firefox"},
-	)
-	return cmds
-}
-
-// Open tries to open url in a browser and reports whether it succeeded.
-func BrowserOpen(url string) bool {
-	for _, args := range BrowserCommands() {
-		cmd := exec.Command(args[0], append(args[1:], url)...)
-		if cmd.Start() == nil && appearsSuccessful(cmd, 3*time.Second) {
-			return true
-		}
-	}
-	return false
-}
-
-// appearsSuccessful reports whether the command appears to have run successfully.
-// If the command runs longer than the timeout, it's deemed successful.
-// If the command runs within the timeout, it's deemed successful if it exited cleanly.
-func appearsSuccessful(cmd *exec.Cmd, timeout time.Duration) bool {
-	errc := make(chan error, 1)
-	go func() {
-		errc <- cmd.Wait()
-	}()
-
-	select {
-	case <-time.After(timeout):
-		return true
-	case err := <-errc:
-		return err == nil
-	}
+	browser.Open(jepsenUrl)
 }
 
 func runJepsenTest(test *JepsenTest) int {
@@ -376,7 +321,7 @@ func main() {
 	}
 	if *web && !*dryRun {
 		// Open Jaeger UI
-		BrowserOpen("http://localhost:16686")
+		browser.Open("http://localhost:16686")
 	}
 
 	workloads := strings.Split(*workload, " ")
